Twitter Bot - Sending User Objects in Find_Followers
Screen names can be changed, so it's important that we have access to the user object so we can use the user's ID.

Since find_followers accepts to search both by screen_name and id (sending a map with the key the same as the search_param), this might lead to a situation where we need to search without having the user object stored in the first place.

As such, now `find_followers` routine now also sends the users objects.
If it's, for some reason, unable to find any of them, it prints out a warning