Twitter Bot - Bugfixing + Better Logging
Due to a bug that went unnoticed, timelines were being read 3 to 5 times (max_depth/max_jumps in read_timeline function).

This was because whenever we jumped users, we compared if the user of the tweet was the same as the user object whose timeline we were reading. If they were different, we'd jump.

Since we didn't implement equality methods, this was always true... (Python does `a is b` if `a.__eq__(b)` isn't defined)

Implemented equality methods, as well as change the comparison to compare based on ids, lest some refactoring in the models be done again and someone forgets to implement the equality methods again, at least comparing by ids makes sure it will keep working (assuming no one changes the attributes ofc)