Twitter Bot - Bugfixing Exception Handling
Due to the way we were using the `log_account_suspended` decorator, it was exiting on the first time an exception occurred in the `run` method.
Thus, removed the decorator and added a helper method in TwitterBot so we don't have to repeat the code, while the bot can keep processing the tasks.